requirements.txt
langchain 
langchain-community 
pydf 
faiss-cpu 
chroma 
langgraph 
langchain-groq
pypdf
sentence-transformers
langchain-huggingface
python-dotenv
streamlit